+++ lib/src/hardware/sensor_manager.dart
@@ -39,6 +39,6 @@ abstract class SensorManager {
     };
     final String channelID =
         await _channel.invokeMethod('registerListener', request);
-    return EventChannel('flutter_android/SensorManager/$channelID');
+    return (channelID != null) ? EventChannel('flutter_android/SensorManager/$channelID') : null;
   }
 }