+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.3.4] - 2019-02-15
 ### Added
 - `android_location.Location` class
+- `android_hardware.Sensor#subscribe()` method
+- `android_hardware.SensorEvent` class
+- `android_hardware.SensorManager.getDefaultSensor()` static method
 
 ## [0.3.3] - 2019-02-15
 ### Added