+++ CHANGELOG.md
@@ -7,7 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.3.5] - 2019-02-15
-### Added
+### Changed
+- Fixed a bug in `android_hardware.SensorManager.registerListener()`
 
 ## [0.3.4] - 2019-02-15
 ### Added